Prevent targeting dead monsters

+++ Source Main 5.2/source/w_CharacterInfo.h
@@ -166,7 +166,7 @@ public:
     BYTE        TargetY;
     BYTE        SkillX;
     BYTE        SkillY;
-    BYTE        Appear;
+    float        Appear;
     BYTE	    CurrentSkill;
     BYTE        CastRenderTime; // unused?
     BYTE        m_byFriend;
